requirements.txt
google-api-python-client>=2.70.0
python-dotenv>=1.0.0
tqdm>=4.64.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
structlog>=22.3.0
flask>=2.0.0
gunicorn>=21.2.0
Pillow>=10.0.0
zipfile36>=0.1.3
html2image>=2.0.0
psycopg2-binary>=2.9.0 